@@ -1,4 +1,4 @@
-name: MedicineReminder
+name: Vitel
 description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -31,18 +31,30 @@ dependencies:
   intl: ^0.17.0
   get: ^4.3.8
   flutter_svg: ^0.22.0
+  firebase_core: ^1.6.0
+  firebase_auth: ^3.1.0
+  google_sign_in: ^5.1.0
+  sqflite: ^2.0.0+4
+  country_code_picker: ^2.0.2
 
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  sms_autofill: ^2.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   
+  flutter_launcher_icons: ^0.9.2
   
-
+  
+flutter_icons:
+  image_path: "assets/logo/app_icon.png"
+  android: true
+  ios: true
+  remove_alpha_ios: true
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -57,7 +69,13 @@ flutter:
 
   # To add assets to your application, add an assets section, like this:
   assets:
-    - assets/google 1.svg
+    - assets/images/google.svg
+    - assets/images/home.svg
+    - assets/images/profile.svg
+    - assets/images/settings.svg
+    - assets/images/no_medicine.svg
+    - assets/images/no_vital.png
+    - data/
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
@@ -70,17 +88,11 @@ flutter:
   # "family" key with the font family name, and a "fonts" key with a
   # list giving the asset and other descriptors for the font. For
   # example:
-  # fonts:
-  #   - family: Schyler
-  #     fonts:
-  #       - asset: fonts/Schyler-Regular.ttf
-  #       - asset: fonts/Schyler-Italic.ttf
-  #         style: italic
-  #   - family: Trajan Pro
-  #     fonts:
-  #       - asset: fonts/TrajanPro.ttf
-  #       - asset: fonts/TrajanPro_Bold.ttf
-  #         weight: 700
+  fonts:
+    - family: Poppins
+      fonts:
+        - asset: assets/fonts/Poppins-SemiBold.ttf
+
   #
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
