@@ -1,10 +1,15 @@
-import 'package:MedicineReminder/view/widgets/dashboard_widget/calendar.dart';
-import 'package:MedicineReminder/view/widgets/dashboard_widget/dashboard_appbar.dart';
-import 'package:MedicineReminder/view/widgets/dashboard_widget/medicine_detail/medicine_detail.dart';
-import 'package:MedicineReminder/view/widgets/dashboard_widget/medicine_info_tile.dart';
-import 'package:MedicineReminder/view/widgets/dashboard_widget/new_medicine_detail/new_medicine_detail.dart';
+import 'package:Vitals/controller/auth_user_controller.dart';
+import 'package:Vitals/controller/get_medicine/get_medicine.dart';
+import 'package:Vitals/main.dart';
+import 'package:Vitals/view/widgets/dashboard_widget/calendar.dart';
+import 'package:Vitals/view/widgets/dashboard_widget/dashboard_appbar.dart';
+import 'package:Vitals/view/widgets/dashboard_widget/medicine_detail/medicine_detail.dart';
+import 'package:Vitals/view/widgets/dashboard_widget/medicine_info_tile.dart';
+import 'package:Vitals/view/widgets/dashboard_widget/new_medicine_detail/new_medicine_detail.dart';
 
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
+import 'package:get/get_state_manager/src/simple/get_state.dart';
 
 class DashBoard extends StatefulWidget {
   int currentIndex;
@@ -15,43 +20,103 @@ class DashBoard extends StatefulWidget {
 }
 
 class _DashBoardState extends State<DashBoard> {
+  AuthUserController authUserController = Get.put(AuthUserController());
+  double displayHeight = 0;
+
   @override
   Widget build(BuildContext context) {
+    displayHeight = MediaQuery.of(context).size.height;
+    String headerTitle;
+
     return Scaffold(
-     backgroundColor:  Colors.blue,
+      backgroundColor: kPrimaryColor,
+      appBar: AppBar(
+        title: GetBuilder<AuthUserController>(
+          init: AuthUserController(),
+          builder: (authController) {
+            if (authController.userName != '' && authController.userPhone == '') {
+              headerTitle = authController.userName;
+            } else {
+              headerTitle = authController.userPhone;
+            }
+            return Center(
+              child: Text('Welcome, ' + headerTitle),
+            );
+          },
+        ),
+        automaticallyImplyLeading: false,
+        backgroundColor: kPrimaryColor,
+        elevation: 0,
+        toolbarHeight: 50,
+      ),
       body: Container(
-      // color:   Color(0xffEDF7FF),
-        padding: EdgeInsets.only(bottom:50),
         child: SingleChildScrollView(
-          child: Column(
-            children: <Widget>[
-              Container(
-                padding: EdgeInsets.only(top: 50,bottom: 10),
-                width: MediaQuery.of(context).size.width,
-                decoration: BoxDecoration(
-                  borderRadius: BorderRadius.all(Radius.circular(5))
-                ),
-                child: Column(
-                  children: <Widget>[
-                    CustomCalendar('week'),
-                  ],
-                ),
-              ),
-             // SizedBox(height:10),
-              InkWell(
-                onTap: medicineDetail,
-                child: Container(
-                  decoration: BoxDecoration(
-                      borderRadius: BorderRadius.only(
-                          topLeft: Radius.circular(30),
-                          topRight: Radius.circular(30)),
-                      color:  Color(0xffEDF7FF),
+          child: GetBuilder<GetMedicineController>(
+            init: GetMedicineController(),
+            builder: (refreshTabController) {
+              return Column(
+                children: <Widget>[
+                  Container(
+                    padding: EdgeInsets.only(top: 0, bottom: 10),
+                    width: MediaQuery.of(context).size.width,
+                    decoration: BoxDecoration(
+                      borderRadius: BorderRadius.all(
+                        Radius.circular(5),
+                      ),
+                    ),
+                    child: CustomCalendar('week'),
                   ),
-                  child: MedicineInfoTile(),
-                ),
-              ),
-              
-            ],
+                  GetBuilder<GetMedicineController>(
+                    init: GetMedicineController(),
+                    builder: (vitalController) {
+                      return vitalController.selectedVital.length > 0
+                          ? Container(
+                              padding: EdgeInsets.only(
+                                  bottom:
+                                      vitalController.selectedVital.length < 3
+                                          ? 300
+                                          : 100),
+                              decoration: BoxDecoration(
+                                borderRadius: BorderRadius.only(
+                                  topLeft: Radius.circular(30),
+                                  topRight: Radius.circular(30),
+                                ),
+                                color: Color(0xffEDF7FF),
+                              ),
+                              child: MedicineInfoTile(),
+                            )
+                          : Container(
+                              height: MediaQuery.of(context).size.height - 280,
+                              width: MediaQuery.of(context).size.width,
+                              decoration: BoxDecoration(
+                                borderRadius: BorderRadius.only(
+                                  topLeft: Radius.circular(30),
+                                  topRight: Radius.circular(30),
+                                ),
+                                color: Color(0xffEDF7FF),
+                              ),
+                              child: Column(
+                                mainAxisAlignment: MainAxisAlignment.center,
+                                children: [
+                                  Image.asset('assets/images/no_vital.png'),
+                                  SizedBox(
+                                    height: 30,
+                                  ),
+                                  Text(
+                                    'No Medicine Added Today',
+                                    style: TextStyle(
+                                      color: kPrimaryColor,
+                                      fontSize: 18,
+                                    ),
+                                  )
+                                ],
+                              ),
+                            );
+                    },
+                  )
+                ],
+              );
+            },
           ),
         ),
       ),
@@ -85,7 +150,14 @@ class _DashBoardState extends State<DashBoard> {
         });
   }
 
-  void medicineDetail() {
+/*   void medicineDetail() {
+    double modalHeight = 0;
+    if (displayHeight < 750) {
+      modalHeight = MediaQuery.of(context).size.height * 0.8;
+    }
+    if (displayHeight >= 750) {
+      modalHeight = modalHeight = MediaQuery.of(context).size.height * 0.72;
+    }
     showModalBottomSheet(
         isScrollControlled: true,
         isDismissible: true,
@@ -93,16 +165,16 @@ class _DashBoardState extends State<DashBoard> {
         context: context,
         builder: (context) {
           return Container(
-            height: MediaQuery.of(context).size.height * 0.78,
+            height: modalHeight,
             decoration: new BoxDecoration(
-              color: Colors.white,
+              color: Color(0xffEDF7FF),
               borderRadius: new BorderRadius.only(
                 topLeft: const Radius.circular(35.0),
                 topRight: const Radius.circular(35.0),
               ),
             ),
-            child: MedicineDetail(),
+            child: MedicineDetail(vital),
           );
         });
-  }
+  } */
 }
