@@ -1,9 +1,15 @@
-import 'package:MedicineReminder/controller/sign_in_controller.dart';
-import 'package:MedicineReminder/view/pages/home/home_page.dart';
+import 'package:Vitals/Authentication/g_auth.dart';
+import 'package:Vitals/controller/auth_user_controller.dart';
+import 'package:Vitals/controller/sign_in_controller.dart';
+import 'package:Vitals/view/pages/home/home_page.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:MedicineReminder/view/widgets/sign_up_widget/sign_up_text_fields.dart';
+import 'package:Vitals/view/widgets/sign_up_widget/sign_up_text_fields.dart';
 import 'package:get/get.dart';
 import 'package:flutter_svg/svg.dart';
+import 'package:Vitals/main.dart';
+
+import '../../../main.dart';
 
 class SignUpPage extends StatefulWidget {
   @override
@@ -12,6 +18,9 @@ class SignUpPage extends StatefulWidget {
 
 class _SignUpPageState extends State<SignUpPage> {
   SignInController signInController = Get.put(SignInController());
+
+  AuthUserController authUserController = Get.put(AuthUserController());
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -27,11 +36,14 @@ class _SignUpPageState extends State<SignUpPage> {
                 children: <Widget>[
                   Text(
                     'Skip',
-                    style: TextStyle(color: Colors.blue, fontSize: 14),
+                    style: TextStyle(
+                        color: kPrimaryColor,
+                        fontSize: 14,
+                        fontWeight: FontWeight.w600),
                   ),
                   Icon(
                     Icons.arrow_right_alt_sharp,
-                    color: Colors.blue,
+                    color: kPrimaryColor,
                     size: 30,
                   )
                 ],
@@ -43,60 +55,83 @@ class _SignUpPageState extends State<SignUpPage> {
       backgroundColor: Color(0xffEDF7FF),
       body: SingleChildScrollView(
         child: Column(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: <Widget>[
-              Center(
-                child: Container(
-                  margin: EdgeInsets.only(top: 30),
-                  child: Text(
-                    'Create Account',
-                    style: TextStyle(
-                        color: Colors.blue,
-                        fontSize: 24,
-                        fontWeight: FontWeight.w700),
-                  ),
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: <Widget>[
+            Align(
+              alignment: Alignment.centerLeft,
+              child: Container(
+                width: MediaQuery.of(context).size.width * .8,
+                margin: EdgeInsets.only(
+                  top: 50,
+                  left: MediaQuery.of(context).size.width * 0.097,
+                ),
+                child: Text(
+                  'Create Account',
+                  style: TextStyle(
+                      color: kPrimaryColor,
+                      fontSize: 24,
+                      fontWeight: FontWeight.w700),
                 ),
               ),
-              Center(
-                child: Container(
-                  //margin: EdgeInsets.only(top:40),
-                  child: Text(
-                    'Sign up to get Started',
-                    style: TextStyle(
-                        color: Color(0xff606365).withOpacity(0.6),
-                        fontSize: 20,
-                        fontWeight: FontWeight.w400),
-                  ),
+            ),
+            Align(
+              alignment: Alignment.centerLeft,
+              child: Container(
+                width: MediaQuery.of(context).size.width * .8,
+                margin: EdgeInsets.only(
+                  top: 2.5,
+                  left: MediaQuery.of(context).size.width * 0.097,
+                ),
+                child: Text(
+                  'Sign In To Get Started!',
+                  style: TextStyle(
+                      color: Color(0xff606365).withOpacity(0.6),
+                      fontSize: 20,
+                      fontWeight: FontWeight.w600),
                 ),
               ),
-              TextFieldWidget(),
-              SizedBox(height: 30),
-              InkWell(
-                onTap: () => Get.find<SignInController>().signedIn(true),
-                child: Container(
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.all(Radius.circular(10)),
-                    color: Colors.blue,
-                  ),
-                  margin: EdgeInsets.only(left: 33, right: 33),
-                  height: 50,
-                  width: MediaQuery.of(context).size.width,
-                  child: Center(
-                      child: Text(
-                    'Sign up',
-                    style: TextStyle(
-                        color: Colors.white,
-                        fontSize: 20,
-                        fontWeight: FontWeight.w600),
-                  )),
-                ),
-              ),
-              SizedBox(height: 20),
-              Container(
+            ),
+            TextFieldWidget(),
+            SizedBox(height: 20),
+            InkWell(
+              onTap: () {
+                googleLogin().then(
+                  (result) {
+                    // ignore: unnecessary_null_comparison
+                    if (result != '') {
+                      Navigator.of(context).pushReplacement(
+                        MaterialPageRoute(
+                          fullscreenDialog: true,
+                          builder: (context) {
+                            User? firebaseUser =
+                                FirebaseAuth.instance.currentUser;
+
+                            // this means google login
+                            String username =
+                                firebaseUser?.displayName.toString() ??
+                                    'Google Username';
+                            String email = firebaseUser?.email.toString() ??
+                                'Google User Email';
+                            String image = firebaseUser?.photoURL.toString() ??
+                                'https://image.pngaaa.com/677/884677-middle.png';
+
+                            Get.find<AuthUserController>()
+                                .updateVal(username, email, '', image);
+
+                            return Dashboard();
+                          },
+                        ),
+                      );
+                    }
+                  },
+                );
+              },
+              child: Container(
                 decoration: BoxDecoration(
                   borderRadius: BorderRadius.all(Radius.circular(10)),
                   border: Border.all(
-                    color: Colors.blue.shade200.withOpacity(0.5),
+                    width: 2,
+                    color: kPrimaryColor,
                   ),
                 ),
                 height: 50,
@@ -107,23 +142,27 @@ class _SignUpPageState extends State<SignUpPage> {
                   children: [
                     Container(
                       child: SvgPicture.asset(
-                        "assets/google 1.svg",
+                        "assets/images/google.svg",
+                        height: 20,
+                        width: 20,
                       ),
                     ),
-                    SizedBox(width:10),
+                    SizedBox(width: 10),
                     Center(
                       child: Text(
                         'Sign up with Google',
                         style: TextStyle(
-                            color: Colors.blue,
-                            fontSize: 20,
+                            color: kPrimaryColor,
+                            fontSize: 18,
                             fontWeight: FontWeight.w600),
                       ),
                     ),
                   ],
                 ),
-              )
-            ]),
+              ),
+            ),
+          ],
+        ),
       ),
     );
   }
