@@ -1,18 +1,30 @@
-import 'package:MedicineReminder/view/widgets/dashboard_widget/new_medicine_detail/meal/after_meal.dart';
-import 'package:MedicineReminder/view/widgets/dashboard_widget/new_medicine_detail/medicine_info_text_field.dart';
-import 'package:MedicineReminder/view/widgets/dashboard_widget/new_medicine_detail/new_medicine_detail.dart';
+import 'package:Vitals/controller/get_medicine/get_medicine.dart';
+import 'package:Vitals/database/vital_reprository.dart';
+import 'package:Vitals/model/medicine_model.dart';
+import 'package:Vitals/view/widgets/dashboard_widget/new_medicine_detail/meal/after_meal.dart';
+import 'package:Vitals/view/widgets/dashboard_widget/new_medicine_detail/meal/before_meal.dart';
+import 'package:Vitals/view/widgets/dashboard_widget/new_medicine_detail/medicine_info_text_field.dart';
+import 'package:Vitals/view/widgets/dashboard_widget/new_medicine_detail/new_medicine_detail.dart';
+import 'package:Vitals/view/widgets/dashboard_widget/medicine_detail/update_medicine.dart';
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
 
 class MedicineDetail extends StatefulWidget {
+  final Vital vital;
+  MedicineDetail(this.vital);
   @override
   _MedicineDetailState createState() => _MedicineDetailState();
 }
 
 class _MedicineDetailState extends State<MedicineDetail> {
-  int dose=3;
+   late double displayHeight;
+  int dose = 3;
   TextEditingController editingController = new TextEditingController();
+  GetMedicineController getmedicineController =
+      Get.put(GetMedicineController());
   @override
   Widget build(BuildContext context) {
+    displayHeight = MediaQuery.of(context).size.height;
     Size size = MediaQuery.of(context).size;
     return Container(
       child: Column(
@@ -30,15 +42,17 @@ class _MedicineDetailState extends State<MedicineDetail> {
                   style: TextStyle(fontSize: 16, color: Colors.grey),
                 )),
                 Container(
-                  child: Text('Indomate',
+                  child: Text(widget.vital.name,
                       style: TextStyle(
                         fontSize: 24,
                         color: Colors.blue,
                         fontWeight: FontWeight.w700,
                       )),
                 ),
-                SizedBox(height:5),
-                AfterMeal('after', 'other')
+                SizedBox(height: 5),
+                widget.vital.afterMeal == 1
+                    ? AfterMeal('after', 'other')
+                    : BeforeMeal('before', 'other')
               ],
             ),
           ),
@@ -59,170 +73,188 @@ class _MedicineDetailState extends State<MedicineDetail> {
                             fontWeight: FontWeight.w700),
                       ),
                       SizedBox(height: 5),
-                       Container(
-            margin: EdgeInsets.only(top: 10),
-            child: Column(
-              children: <Widget>[
-                Container(
-                  child: Row(
-                    children: <Widget>[
-                      dose >= 1
-                          ? InkWell(
-                              onTap: () {
-                                //_showDialog('dose1');
-                              },
-                              child: Container(
-                                width: 95,
-                                height: 22,
-                                decoration: BoxDecoration(
-                                  borderRadius: new BorderRadius.all(
-                                    const Radius.circular(5.0),
-                                  ),
-                                  color: Colors.blue,
-                                ),
-                                margin: EdgeInsets.only(right: 5),
-                                child: Center(
-                                  child: Text(
-                                    'dose time',
-                                    style: TextStyle(fontSize: 14,color: Colors.white),
-                                  ),
-                                ),
+                      Container(
+                        margin: EdgeInsets.only(top: 10),
+                        child: Column(
+                          children: <Widget>[
+                            Container(
+                              child: Row(
+                                children: <Widget>[
+                                  widget.vital.doseOne.isNotEmpty
+                                      ? InkWell(
+                                          onTap: () {
+                                            //_showDialog('dose1');
+                                          },
+                                          child: Container(
+                                            width: 95,
+                                            height: 22,
+                                            decoration: BoxDecoration(
+                                              borderRadius:
+                                                  new BorderRadius.all(
+                                                const Radius.circular(5.0),
+                                              ),
+                                              color: Colors.blue,
+                                            ),
+                                            margin: EdgeInsets.only(right: 5),
+                                            child: Center(
+                                              child: Text(
+                                                widget.vital.doseOne,
+                                                style: TextStyle(
+                                                    fontSize: 14,
+                                                    color: Colors.white),
+                                              ),
+                                            ),
+                                          ),
+                                        )
+                                      : SizedBox(),
+                                  widget.vital.doseTwo.isNotEmpty
+                                      ? InkWell(
+                                          onTap: () {
+                                            // _showDialog( 'dose2');
+                                          },
+                                          child: Container(
+                                            width: 95,
+                                            height: 22,
+                                            decoration: BoxDecoration(
+                                              borderRadius:
+                                                  new BorderRadius.all(
+                                                const Radius.circular(5.0),
+                                              ),
+                                              color: Colors.blue,
+                                            ),
+                                            margin: EdgeInsets.only(right: 5),
+                                            child: Center(
+                                              child: Text(
+                                                widget.vital.doseTwo,
+                                                style: TextStyle(
+                                                    fontSize: 14,
+                                                    color: Colors.white),
+                                              ),
+                                            ),
+                                          ),
+                                        )
+                                      : SizedBox(),
+                                  widget.vital.doseThree.isNotEmpty
+                                      ? InkWell(
+                                          onTap: () {
+                                            //_showDialog('dose3');
+                                          },
+                                          child: Container(
+                                            width: 95,
+                                            height: 22,
+                                            decoration: BoxDecoration(
+                                              borderRadius:
+                                                  new BorderRadius.all(
+                                                const Radius.circular(5.0),
+                                              ),
+                                              color: Colors.blue,
+                                            ),
+                                            margin: EdgeInsets.only(right: 5),
+                                            child: Center(
+                                              child: Text(
+                                                widget.vital.doseThree,
+                                                style: TextStyle(
+                                                    fontSize: 14,
+                                                    color: Colors.white),
+                                              ),
+                                            ),
+                                          ),
+                                        )
+                                      : SizedBox(),
+                                ],
                               ),
-                            )
-                          : SizedBox(),
-                      dose >= 2
-                          ? InkWell(
-                              onTap: () {
-                               // _showDialog( 'dose2');
-                              },
-                              child: Container(
-                                width: 95,
-                                height: 22,
-                                decoration: BoxDecoration(
-                                  borderRadius: new BorderRadius.all(
-                                    const Radius.circular(5.0),
-                                  ),
-                                  color: Colors.blue,
-                                ),
-                                margin: EdgeInsets.only(right: 5),
-                                child: Center(
-                                  child: Text(
-                                    'dose time',
-                                    style: TextStyle(fontSize: 14,color: Colors.white),
-                                  ),
-                                ),
+                            ),
+                            Container(
+                              margin: EdgeInsets.only(top: 10),
+                              child: Row(
+                                children: <Widget>[
+                                  widget.vital.doseFour.isNotEmpty
+                                      ? InkWell(
+                                          onTap: () {
+                                            // _showDialog('dose4');
+                                          },
+                                          child: Container(
+                                            width: 95,
+                                            height: 22,
+                                            decoration: BoxDecoration(
+                                              borderRadius:
+                                                  new BorderRadius.all(
+                                                const Radius.circular(5.0),
+                                              ),
+                                              color: Colors.blue,
+                                            ),
+                                            margin: EdgeInsets.only(right: 5),
+                                            child: Center(
+                                              child: Text(
+                                                widget.vital.doseFour,
+                                                style: TextStyle(
+                                                    fontSize: 14,
+                                                    color: Colors.white),
+                                              ),
+                                            ),
+                                          ),
+                                        )
+                                      : SizedBox(),
+                                  widget.vital.doseFive.isNotEmpty
+                                      ? InkWell(
+                                          onTap: () {
+                                            //_showDialog('dose5');
+                                          },
+                                          child: Container(
+                                            width: 95,
+                                            height: 22,
+                                            decoration: BoxDecoration(
+                                              borderRadius:
+                                                  new BorderRadius.all(
+                                                const Radius.circular(5.0),
+                                              ),
+                                              color: Colors.blue,
+                                            ),
+                                            margin: EdgeInsets.only(right: 5),
+                                            child: Center(
+                                              child: Text(
+                                                widget.vital.doseFive,
+                                                style: TextStyle(
+                                                    fontSize: 14,
+                                                    color: Colors.white),
+                                              ),
+                                            ),
+                                          ),
+                                        )
+                                      : SizedBox(),
+                                  widget.vital.doseSix.isNotEmpty
+                                      ? InkWell(
+                                          onTap: () {
+                                            //_showDialog('dose6');
+                                          },
+                                          child: Container(
+                                            width: 95,
+                                            height: 22,
+                                            decoration: BoxDecoration(
+                                              borderRadius:
+                                                  new BorderRadius.all(
+                                                const Radius.circular(5.0),
+                                              ),
+                                              color: Colors.blue,
+                                            ),
+                                            margin: EdgeInsets.only(right: 5),
+                                            child: Center(
+                                              child: Text(
+                                                widget.vital.doseSix,
+                                                style: TextStyle(
+                                                    fontSize: 14,
+                                                    color: Colors.white),
+                                              ),
+                                            ),
+                                          ),
+                                        )
+                                      : SizedBox(),
+                                ],
                               ),
-                            )
-                          : SizedBox(),
-                      dose >= 3
-                          ? InkWell(
-                              onTap: () {
-                                //_showDialog('dose3');
-                              },
-                              child: Container(
-                                width: 95,
-                                height: 22,
-                                decoration: BoxDecoration(
-                                  borderRadius: new BorderRadius.all(
-                                    const Radius.circular(5.0),
-                                  ),
-                                  color: Colors.blue,
-                                ),
-                                margin: EdgeInsets.only(right: 5),
-                                child: Center(
-                                  child: Text(
-                                    'dose time',
-                                    style: TextStyle(fontSize: 14,color: Colors.white),
-                                  ),
-                                ),
-                              ),
-                            )
-                          : SizedBox(),
-                    ],
-                  ),
-                ),
-                Container(
-                  margin: EdgeInsets.only(top: 10),
-                  child: Row(
-                    children: <Widget>[
-                      dose >= 4
-                          ? InkWell(
-                              onTap: () {
-                               // _showDialog('dose4');
-                              },
-                              child: Container(
-                                width: 95,
-                                height: 22,
-                                decoration: BoxDecoration(
-                                  borderRadius: new BorderRadius.all(
-                                    const Radius.circular(5.0),
-                                  ),
-                                  color: Colors.blue,
-                                ),
-                                margin: EdgeInsets.only(right: 5),
-                                child: Center(
-                                  child: Text(
-                                    'dose time',
-                                    style: TextStyle(fontSize: 14,color: Colors.white),
-                                  ),
-                                ),
-                              ),
-                            )
-                          : SizedBox(),
-                      dose >= 5
-                          ? InkWell(
-                              onTap: () {
-                                //_showDialog('dose5');
-                              },
-                              child: Container(
-                                width: 95,
-                                height: 22,
-                                decoration: BoxDecoration(
-                                  borderRadius: new BorderRadius.all(
-                                    const Radius.circular(5.0),
-                                  ),
-                                  color: Colors.blue,
-                                ),
-                                margin: EdgeInsets.only(right: 5),
-                                child: Center(
-                                  child: Text(
-                                    'dose time',
-                                    style: TextStyle(fontSize: 14,color: Colors.white),
-                                  ),
-                                ),
-                              ),
-                            )
-                          : SizedBox(),
-                      dose == 6
-                          ? InkWell(
-                              onTap: () {
-                                //_showDialog('dose6');
-                              },
-                              child: Container(
-                                width: 95,
-                                height: 22,
-                                decoration: BoxDecoration(
-                                  borderRadius: new BorderRadius.all(
-                                    const Radius.circular(5.0),
-                                  ),
-                                  color: Colors.blue,
-                                ),
-                                margin: EdgeInsets.only(right: 5),
-                                child: Center(
-                                  child: Text(
-                                    'dose time',
-                                    style: TextStyle(fontSize: 14,color: Colors.white),
-                                  ),
-                                ),
-                              ),
-                            )
-                          : SizedBox(),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          )
+                            ),
+                          ],
+                        ),
+                      )
                     ],
                   ),
                 ),
@@ -244,7 +276,11 @@ class _MedicineDetailState extends State<MedicineDetail> {
                           children: [
                             Container(
                               child: Text(
-                                'Total 30 Days ' + '|' + " ",
+                                'Total ' +
+                                    widget.vital.program.toString() +
+                                    ' Days ' +
+                                    '|' +
+                                    " ",
                                 style: TextStyle(
                                     color: Colors.grey,
                                     fontWeight: FontWeight.w600),
@@ -252,7 +288,7 @@ class _MedicineDetailState extends State<MedicineDetail> {
                             ),
                             Container(
                               child: Text(
-                                '12 Days Left',
+                                '_ Days Left',
                                 style: TextStyle(
                                     color: Colors.grey,
                                     fontWeight: FontWeight.w600,
@@ -283,7 +319,11 @@ class _MedicineDetailState extends State<MedicineDetail> {
                           children: [
                             Container(
                               child: Text(
-                                'Total 90 Pills ' + '|' + " ",
+                                'Total ' +
+                                    widget.vital.quantity.toString() +
+                                    ' Pills ' +
+                                    '|' +
+                                    " ",
                                 style: TextStyle(
                                     color: Colors.grey,
                                     fontWeight: FontWeight.w600),
@@ -291,7 +331,7 @@ class _MedicineDetailState extends State<MedicineDetail> {
                             ),
                             Container(
                               child: Text(
-                                '36 Pills Left',
+                                '_ Pills Left',
                                 style: TextStyle(
                                     color: Colors.grey,
                                     fontWeight: FontWeight.w600,
@@ -304,10 +344,16 @@ class _MedicineDetailState extends State<MedicineDetail> {
                     ],
                   ),
                 ),
-                SizedBox(height:30),
-                 Center(
+                SizedBox(height: 30),
+                Center(
                   child: InkWell(
-                    onTap: null,
+                    onTap: () {
+                      /* await Repository.deleteData('Vitals', widget.vital.id);
+                      getmedicineController.getAllVitalFromDb();
+                      Navigator.pop(context); */
+                      Navigator.pop(context);
+                      deletAlert(widget.vital, context);
+                    },
                     child: Container(
                       width: 115,
                       height: 30,
@@ -315,11 +361,16 @@ class _MedicineDetailState extends State<MedicineDetail> {
                         borderRadius: BorderRadius.all(Radius.circular(10.0)),
                         color: Colors.red,
                       ),
-                      child: Center(child: Text('Delete',style: TextStyle(fontSize:14,color: Colors.white),),),
+                      child: Center(
+                        child: Text(
+                          'Delete',
+                          style: TextStyle(fontSize: 14, color: Colors.white),
+                        ),
+                      ),
                     ),
                   ),
                 ),
-                SizedBox(height:27),
+                SizedBox(height: 27),
                 Center(
                   child: InkWell(
                     onTap: addMedicine,
@@ -330,7 +381,12 @@ class _MedicineDetailState extends State<MedicineDetail> {
                         borderRadius: BorderRadius.all(Radius.circular(10.0)),
                         color: Colors.blue,
                       ),
-                      child: Center(child: Text('Edit Schedule',style: TextStyle(fontSize: 20,color: Colors.white),),),
+                      child: Center(
+                        child: Text(
+                          'Edit Schedule',
+                          style: TextStyle(fontSize: 20, color: Colors.white),
+                        ),
+                      ),
                     ),
                   ),
                 )
@@ -343,6 +399,20 @@ class _MedicineDetailState extends State<MedicineDetail> {
   }
 
   void addMedicine() {
+
+    late double modalHeight;
+    if (displayHeight < 700) {
+      modalHeight = MediaQuery.of(context).size.height * 0.9;
+    }
+    if (displayHeight > 700 && displayHeight < 750) {
+      modalHeight = modalHeight = MediaQuery.of(context).size.height * 0.87;
+      //modalHeight = modalHeight = MediaQuery.of(context).size.height * 0.7;
+    }
+    if (displayHeight > 750) {
+      modalHeight = modalHeight = MediaQuery.of(context).size.height * 0.75;
+      //modalHeight = modalHeight = MediaQuery.of(context).size.height * 0.7;
+      // responsive modal implementation
+    }
     showModalBottomSheet(
         isScrollControlled: true,
         isDismissible: true,
@@ -350,7 +420,7 @@ class _MedicineDetailState extends State<MedicineDetail> {
         context: context,
         builder: (context) {
           return Container(
-            height: MediaQuery.of(context).size.height * 0.78,
+            height: modalHeight,
             decoration: new BoxDecoration(
               color: Color(0xffEDF7FF),
               borderRadius: new BorderRadius.only(
@@ -358,8 +428,72 @@ class _MedicineDetailState extends State<MedicineDetail> {
                 topRight: const Radius.circular(35.0),
               ),
             ),
-            child: NewMedicineDetail(),
+            child: UpdateMedicineDetail(widget.vital),
           );
         });
+         
+  }
+
+  void deletAlert(Vital vital, context) {
+    showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: Row(
+            children: [
+              Text('Delete', style: TextStyle(color: Colors.red)),
+              SizedBox(
+                width: 10,
+              ),
+              Text(vital.name),
+            ],
+          ),
+          content: Text("Do you want to delte" + ' ' + vital.name),
+          actions: [
+            Card(
+              color: Colors.green,
+              child: Container(
+                width: 50,
+                height: 30,
+                child: Center(
+                  child: InkWell(
+                    onTap: () {
+                      Navigator.pop(context);
+                    },
+                    child: Text('NO', style: TextStyle(color: Colors.white)),
+                  ),
+                ),
+              ),
+            ),
+            SizedBox(
+              width: 10,
+            ),
+            Card(
+              color: Colors.red,
+              child: Container(
+                width: 50,
+                height: 30,
+                child: Center(
+                  child: InkWell(
+                    onTap: () async {
+                      await Repository.deleteData('Vitals', vital.id);
+                      getmedicineController.getAllVitalFromDb();
+                      Navigator.pop(context);
+                    },
+                    child: Text(
+                      'Yes',
+                      style: TextStyle(color: Colors.white),
+                    ),
+                  ),
+                ),
+              ),
+            ),
+            SizedBox(
+              width: 10,
+            )
+          ],
+        );
+      },
+    );
   }
 }
