@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 class AfterMeal extends StatefulWidget {
   String meal;
   String from;
-  AfterMeal(this.meal,this.from);
+  AfterMeal(this.meal, this.from);
   @override
   _AfterMealState createState() => _AfterMealState();
 }
@@ -12,16 +12,20 @@ class _AfterMealState extends State<AfterMeal> {
   @override
   Widget build(BuildContext context) {
     return Container(
-      height: widget.from=='dash'?20:30,
-      width:widget.from=='dash'?100:MediaQuery.of(context).size.width*.35,
+      height: widget.from == 'dash' ? 20 : 30,
+      width:
+          widget.from == 'dash' ? 100 : MediaQuery.of(context).size.width * .35,
       decoration: BoxDecoration(
-        color:widget.meal=='after'? Color(0xffb60CC6B):Colors.grey,
-        borderRadius: BorderRadius.only(
-          topRight:Radius.circular(10),
-          bottomLeft:Radius.circular(10)
-        )
+          color:
+              widget.meal == 'after' ? Color(0xffb60CC6B) : Color(0xFFD3DAD4),
+          borderRadius: BorderRadius.only(
+              topRight: Radius.circular(10), bottomLeft: Radius.circular(10))),
+      child: Center(
+        child: Text(
+          'After Meal',
+          style:widget.from == 'dash' ? TextStyle(fontSize: 14, color: Colors.white):TextStyle(fontSize: 18, color: Colors.white)
+        ),
       ),
-      child: Center(child: Text('After Meal',style: TextStyle(fontSize: 16,color: Colors.white),),),
     );
   }
 }
