@@ -1,8 +1,11 @@
+import 'package:Vitals/controller/add_medicine/add_medicine_controller.dart';
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
 
 class MedicineInfoTextField extends StatefulWidget {
   String type;
-  MedicineInfoTextField({this.type});
+  final String vitalName;
+  MedicineInfoTextField({required this.type,required this.vitalName});
 
   @override
   _MedicineInfoTextFieldState createState() => _MedicineInfoTextFieldState();
@@ -10,11 +13,14 @@ class MedicineInfoTextField extends StatefulWidget {
 
 class _MedicineInfoTextFieldState extends State<MedicineInfoTextField> {
   TextEditingController editingController = new TextEditingController();
+  AddMedicineController addmedicineController =
+      Get.put(AddMedicineController());
   @override
   Widget build(BuildContext context) {
     return Container(
       margin: EdgeInsets.only(
-        left: 40,right: 30
+        left: 40,
+        right: 30,
       ),
       child: TextField(
         onEditingComplete: () {
@@ -22,9 +28,11 @@ class _MedicineInfoTextFieldState extends State<MedicineInfoTextField> {
         },
         onChanged: (text) {
           print(text.length);
+          Get.find<AddMedicineController>().changeName(text);
         },
         onSubmitted: (text) {
           print(text.length);
+          Get.find<AddMedicineController>().changeName(text);
         },
         enabled: true,
         minLines: 1,
@@ -36,15 +44,15 @@ class _MedicineInfoTextFieldState extends State<MedicineInfoTextField> {
         cursorHeight: 20,
         controller: editingController,
         style: TextStyle(
-          fontFamily: "AvenirNextCyr",
+          fontFamily: 'Poppins',
           color: Colors.black,
           fontSize: 18,
         ),
         decoration: InputDecoration(
-          hintText: widget.type,
+          hintText:widget.vitalName==''? widget.type:widget.vitalName,
           hintStyle: TextStyle(
-            fontFamily: "AvenirNextCyr",
-            color: Colors.black,
+            fontFamily: 'Poppins',
+            color: Colors.grey,
             fontSize: 18,
           ),
           counterText: '',
