@@ -1,5 +1,8 @@
+import 'package:Vitals/controller/add_medicine/add_medicine_controller.dart';
+import 'package:Vitals/main.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
 import 'package:numberpicker/numberpicker.dart';
 
 class MedicinePrograme extends StatefulWidget {
@@ -10,6 +13,8 @@ class MedicinePrograme extends StatefulWidget {
 class _MedicineProgrameState extends State<MedicinePrograme> {
   int selectedDays = 0;
   int _currentIntValue = 0;
+  AddMedicineController addmedicineController =
+      Get.put(AddMedicineController());
   @override
   Widget build(BuildContext context) {
     return Container(
@@ -17,34 +22,43 @@ class _MedicineProgrameState extends State<MedicinePrograme> {
           crossAxisAlignment: CrossAxisAlignment.start,
           children: <Widget>[
             Container(
-              margin: EdgeInsets.only(left: 40, top: 12),
+              margin: EdgeInsets.only(left: 40, top: 0),
               child: Text(
                 'Program',
                 style: TextStyle(
                     fontSize: 20,
-                    color: Colors.blue,
+                    color: kPrimaryColor,
                     fontWeight: FontWeight.w600),
               ),
             ),
             SizedBox(height: 10),
             InkWell(
-              onTap: _showDialog,
+              onTap: () {
+                FocusScopeNode currentFocus = FocusScope.of(context);
+
+                if (!currentFocus.hasPrimaryFocus) {
+                  currentFocus.unfocus();
+                }
+                _showDialog();
+              },
               child: Container(
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
                     Container(
-                        margin: EdgeInsets.only(left: 40),
-                        child: Text(
-                          selectedDays == 0 ? 'Ex 30' : selectedDays.toString(),
-                          style: TextStyle(
-                              color: selectedDays == 0
-                                  ? Colors.grey
-                                  : Colors.black,
-                              fontSize: 16),
-                        )),
+                      margin: EdgeInsets.only(left: 40),
+                      child: Text(
+                        selectedDays == 0
+                            ? 'How Many Days?'
+                            : selectedDays.toString(),
+                        style: TextStyle(
+                            color:
+                                selectedDays == 0 ? Colors.grey : Colors.black,
+                            fontSize: 16),
+                      ),
+                    ),
                     Container(
-                      margin: EdgeInsets.only(left: 40, right: 30),
+                      margin: EdgeInsets.only(left: 40, right: 30, top: 10),
                       height: 1.3,
                       width: MediaQuery.of(context).size.width,
                       color: Colors.grey,
@@ -66,117 +80,131 @@ class _MedicineProgrameState extends State<MedicinePrograme> {
         return Dialog(
           backgroundColor: Colors.transparent,
           child: Container(
-            padding: EdgeInsets.only(top: 20, right: 15),
             decoration: BoxDecoration(
               borderRadius: new BorderRadius.all(
                 const Radius.circular(50.0),
               ),
-              color: Colors.blue,
+              color: kPrimaryColor,
             ),
-            height: 280,
-            width: 300,
-            child: Stack(
-              children: <Widget>[
-                Positioned(
-                  left: 90,
-                  bottom: 70,
-                  child: Container(
-                      color: Colors.blue,
-                      height: 180,
-                      width: 100,
-                      child: CupertinoPicker(
+            height: 250,
+            width: 280,
+            child: Center(
+              child: Stack(
+                children: <Widget>[
+                  Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    crossAxisAlignment: CrossAxisAlignment.center,
+                    children: <Widget>[
+                      Container(
+                        padding: EdgeInsets.only(bottom: 10),
+                        color: kPrimaryColor,
+                        height: 150,
+                        width: 290,
+                        child: CupertinoPicker(
                           useMagnifier: false,
-                          looping: false,
+                          selectionOverlay: null,
+                          looping: true,
                           itemExtent: 100,
-                          diameterRatio: 2,
+                          scrollController: FixedExtentScrollController(),
                           onSelectedItemChanged: (int index) {
-                            print(index);
                             setState(() {
                               selectedDays = index;
                             });
+                            Get.find<AddMedicineController>()
+                                .changeProgram(selectedDays);
                           },
                           children: [
                             for (int i = 0; i <= 100; i++)
-                              Container(
-                                color: Colors.blue,
-                                child: Center(
-                                  child: Text(i.toString(),
-                                      style: TextStyle(
-                                          color: Colors.white,
-                                          fontSize: 50,
-                                          fontWeight: FontWeight.w600)),
+                              Center(
+                                child: Container(
+                                  height: 50,
+                                  color: kPrimaryColor,
+                                  child: Text(
+                                    i.toString(),
+                                    style: TextStyle(
+                                        color: Colors.white,
+                                        fontSize: 50,
+                                        fontWeight: FontWeight.w600),
+                                  ),
                                 ),
                               )
-                          ])),
-                ),
-                Positioned(
-                  top: 80,
-                  left: 30,
-                  child: Container(
-                    child: Text(
-                      'Total',
-                      style: TextStyle(
-                          fontSize: 25,
-                          color: Colors.white,
-                          fontWeight: FontWeight.w600),
-                    ),
+                          ],
+                        ),
+                      ),
+                    ],
                   ),
-                ),
-                Positioned(
-                    top: 80,
-                    left: 190,
+                  Positioned(
+                    top: 105,
+                    left: 32,
                     child: Container(
                       child: Text(
-                        'Days',
+                        'Total',
                         style: TextStyle(
                             fontSize: 25,
                             color: Colors.white,
                             fontWeight: FontWeight.w600),
                       ),
-                    )),
-                Positioned(
-                  bottom: 20,
-                  left: 70,
-                  child: InkWell(
-                    onTap: () {
-                      Navigator.pop(context);
-                    },
+                    ),
+                  ),
+                  Positioned(
+                    top: 105,
+                    right: 32,
                     child: Container(
-                      width: 150,
-                      height: 30,
-                      decoration: BoxDecoration(
-                          borderRadius: BorderRadius.all(
-                            Radius.circular(10),
-                          ),
-                          color: Colors.white),
                       child: Center(
                         child: Text(
-                          'Set Programs',
+                          'Days',
                           style: TextStyle(
-                              fontSize: 16,
-                              color: Colors.blue,
+                              fontSize: 25,
+                              color: Colors.white,
                               fontWeight: FontWeight.w600),
                         ),
                       ),
                     ),
                   ),
-                ),
-                Positioned(
-                  top:5,
-                  right:10,
-                  child: InkWell(
-                    onTap: (){
-                      setState(() {
-                        selectedDays=0;
-                      });
-                       Navigator.pop(context);
-                    },
-                                      child: Container(
-                      child: Icon(Icons.close, color: Colors.red[200]),
+                  Positioned(
+                    bottom: 20,
+                    right: 65,
+                    child: InkWell(
+                      onTap: () {
+                        Navigator.pop(context);
+                      },
+                      child: Container(
+                        width: 150,
+                        height: 30,
+                        decoration: BoxDecoration(
+                            borderRadius: BorderRadius.all(
+                              Radius.circular(50),
+                            ),
+                            color: Colors.white),
+                        child: Center(
+                          child: Text(
+                            'Set Program',
+                            style: TextStyle(
+                                fontSize: 16,
+                                color: kPrimaryColor,
+                                fontWeight: FontWeight.w600),
+                          ),
+                        ),
+                      ),
                     ),
                   ),
-                )
-              ],
+                  Positioned(
+                    top: 20,
+                    right: 20,
+                    child: InkWell(
+                      onTap: () {
+                        setState(() {
+                          selectedDays = 0;
+                        });
+                        Navigator.pop(context);
+                      },
+                      child: Container(
+                        child: Icon(Icons.close, color: Colors.white),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
             ),
           ),
         );
