@@ -1,5 +1,8 @@
+import 'package:Vitals/controller/add_medicine/add_medicine_controller.dart';
+import 'package:Vitals/main.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
 
 class MedicineQuantity extends StatefulWidget {
   @override
@@ -8,6 +11,9 @@ class MedicineQuantity extends StatefulWidget {
 
 class _MedicineQuantityState extends State<MedicineQuantity> {
   int selectedQuantity = 0;
+    AddMedicineController addmedicineController =
+      Get.put(AddMedicineController());
+
   @override
   Widget build(BuildContext context) {
     return Container(
@@ -20,13 +26,20 @@ class _MedicineQuantityState extends State<MedicineQuantity> {
                 'Quantity',
                 style: TextStyle(
                     fontSize: 20,
-                    color: Colors.blue,
+                    color: kPrimaryColor,
                     fontWeight: FontWeight.w600),
               ),
             ),
             SizedBox(height: 10),
             InkWell(
-              onTap: _showDialog,
+              onTap:(){
+                FocusScopeNode currentFocus = FocusScope.of(context);
+
+                if (!currentFocus.hasPrimaryFocus) {
+                  currentFocus.unfocus();
+                }
+                _showDialog();
+              } ,
               child: Container(
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
@@ -34,7 +47,9 @@ class _MedicineQuantityState extends State<MedicineQuantity> {
                     Container(
                         margin: EdgeInsets.only(left: 40),
                         child: Text(
-                          selectedQuantity == 0 ? 'Ex 30' : selectedQuantity.toString(),
+                          selectedQuantity == 0
+                              ? 'Ex. 30'
+                              : selectedQuantity.toString(),
                           style: TextStyle(
                               color: selectedQuantity == 0
                                   ? Colors.grey
@@ -42,7 +57,11 @@ class _MedicineQuantityState extends State<MedicineQuantity> {
                               fontSize: 16),
                         )),
                     Container(
-                      margin: EdgeInsets.only(left: 40, right: 30),
+                      margin: EdgeInsets.only(
+                        left: 40,
+                        right: 30,
+                        top: 10,
+                      ),
                       height: 1.3,
                       width: MediaQuery.of(context).size.width,
                       color: Colors.grey,
@@ -55,7 +74,7 @@ class _MedicineQuantityState extends State<MedicineQuantity> {
     );
   }
 
-  void _showDialog() {
+    void _showDialog() {
     // flutter defined function
     showDialog(
       context: context,
@@ -64,115 +83,131 @@ class _MedicineQuantityState extends State<MedicineQuantity> {
         return Dialog(
           backgroundColor: Colors.transparent,
           child: Container(
-            padding: EdgeInsets.only(top: 20, right: 15),
             decoration: BoxDecoration(
               borderRadius: new BorderRadius.all(
                 const Radius.circular(50.0),
               ),
-              color: Colors.blue,
+              color: kPrimaryColor,
             ),
-            height: 280,
-            width: 300,
-            child: Stack(children: <Widget>[
-              Positioned(
-                left: 90,
-                bottom: 70,
-                child: Container(
-                    color: Colors.blue,
-                    height: 180,
-                    width: 100,
-                    child: CupertinoPicker(
-                        itemExtent: 100,
-                        looping: true,
-                        diameterRatio: 2,
-                        useMagnifier: true,
-                        onSelectedItemChanged: (int index) {
-                          print(index);
-                          setState(() {
-                            selectedQuantity = index;
-                          });
-                        },
-                        children: [
-                          for (int i = 0; i <= 100; i++)
-                            Container(
-                              child: Center(
-                                child: Text(i.toString(),
+            height: 250,
+            width: 280,
+            child: Center(
+              child: Stack(
+                children: <Widget>[
+                  Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    crossAxisAlignment: CrossAxisAlignment.center,
+                    children: <Widget>[
+                      Container(
+                        padding: EdgeInsets.only(bottom:10),
+                        color: kPrimaryColor,
+                        height: 150,
+                        width: 290,
+                        child: CupertinoPicker(
+                          useMagnifier: false,
+                          selectionOverlay: null,
+                          looping: true,
+                          itemExtent: 100,
+                          scrollController: FixedExtentScrollController(),
+                          onSelectedItemChanged: (int index) {
+                            setState(() {
+                              selectedQuantity = index;
+                            });
+                             Get.find<AddMedicineController>().changeQuantity(selectedQuantity);
+                          },
+                          children: [
+                            for (int i = 0; i <= 100; i++)
+                              Center(
+                                child: Container(
+                                  height: 50,
+                                  color: kPrimaryColor,
+                                  child: Text(
+                                    i.toString(),
                                     style: TextStyle(
                                         color: Colors.white,
                                         fontSize: 50,
-                                        fontWeight: FontWeight.w600)),
-                              ),
-                            )
-                        ])),
-              ),
-              Positioned(
-                top: 80,
-                left: 30,
-                child: Container(
-                  child: Text(
-                    'Total',
-                    style: TextStyle(
-                        fontSize: 25,
-                        color: Colors.white,
-                        fontWeight: FontWeight.w600),
-                  ),
-                ),
-              ),
-              Positioned(
-                  top: 80,
-                  left: 190,
-                  child: Container(
-                    child: Text(
-                      'Pieces',
-                      style: TextStyle(
-                          fontSize: 25,
-                          color: Colors.white,
-                          fontWeight: FontWeight.w600),
-                    ),
-                  )),
-              Positioned(
-                bottom: 20,
-                left: 70,
-                child: InkWell(
-                  onTap: () {
-                    Navigator.pop(context);
-                  },
-                  child: Container(
-                    width: 150,
-                    height: 30,
-                    decoration: BoxDecoration(
-                        borderRadius: BorderRadius.all(
-                          Radius.circular(10),
+                                        fontWeight: FontWeight.w600),
+                                  ),
+                                ),
+                              )
+                          ],
                         ),
-                        color: Colors.white),
-                    child: Center(
+                      ),
+                    ],
+                  ),
+                  Positioned(
+                    top: 105,
+                    left: 32,
+                    child: Container(
                       child: Text(
-                        'Set Quantity',
+                        'Total',
                         style: TextStyle(
-                            fontSize: 16,
-                            color: Colors.blue,
+                            fontSize: 25,
+                            color: Colors.white,
                             fontWeight: FontWeight.w600),
                       ),
                     ),
                   ),
-                ),
-              ),
-              Positioned(
-                  top:5,
-                  right:10,
-                  child: InkWell(
-                    onTap: (){
-                      setState(() {
-                        selectedQuantity=0;
-                      });
-                       Navigator.pop(context);
-                    },
-                                      child: Container(
-                      child: Icon(Icons.close, color: Colors.red[200]),
+                  Positioned(
+                    top: 105,
+                    right: 15,
+                    child: Container(
+                      child: Center(
+                        child: Text(
+                          'Pieces',
+                          style: TextStyle(
+                              fontSize: 25,
+                              color: Colors.white,
+                              fontWeight: FontWeight.w600),
+                        ),
+                      ),
                     ),
                   ),
-                )
-            ]),
+                  Positioned(
+                    bottom: 20,
+                    right: 65,
+                    child: InkWell(
+                      onTap: () {
+                        Navigator.pop(context);
+                      },
+                      child: Container(
+                        width: 150,
+                        height: 30,
+                        decoration: BoxDecoration(
+                            borderRadius: BorderRadius.all(
+                              Radius.circular(50),
+                            ),
+                            color: Colors.white),
+                        child: Center(
+                          child: Text(
+                            'Set Quantity',
+                            style: TextStyle(
+                                fontSize: 16,
+                                color: kPrimaryColor,
+                                fontWeight: FontWeight.w600),
+                          ),
+                        ),
+                      ),
+                    ),
+                  ),
+                  Positioned(
+                    top: 20,
+                    right: 20,
+                    child: InkWell(
+                      onTap: () {
+                        setState(() {
+                          selectedQuantity = 0;
+                        });
+                        Navigator.pop(context);
+                      },
+                      child: Container(
+                        child: Icon(Icons.close, color: Colors.white),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
           ),
         );
       },
